Mayor Update
We clean recursive calls adding foward declaration
Improve ECS
now an entity can get n component of the same type.
Move templates inside and inl file to get a compile weight

+++ include/Utils/EasingFunctions.h
@@ -14,17 +14,17 @@ namespace MAD::Easings
 
 	inline float easeInSine(float t)
 	{
-		return 1 - std::cos((t * M_PI) / 2);
+		return 1.0f - static_cast<float>(std::cos((t * M_PI) / 2.0f));
 	}
 
 	inline float easeOutSine(float t)
 	{
-		return std::sin((t * M_PI) / 2);
+		return static_cast<float>(std::sin((t * M_PI) / 2.0f));
 	}
 
 	inline float easeInOutSine(float t)
 	{
-		return -(std::cos(M_PI * t) - 1) / 2;
+		return -(static_cast<float>(std::cos(M_PI * t)) - 1.0f) / 2.0f;
 	}
 
 	inline float easeInQuad(float t)
@@ -34,15 +34,15 @@ namespace MAD::Easings
 
 	inline float easeOutQuad(float t)
 	{
-		return 1 - (1 - t) * (1 - t);
+		return 1.0f - (1.0f - t) * (1.0f - t);
 	}
 
 	inline float easeInOutQuad(float t)
 	{
 		if (t < 0.5f)
-			return 2 * t * t;
+			return 2.0f * t * t;
 		else
-			return 1 - std::pow(-2 * t + 2, 2) / 2;
+			return 1.0f - std::pow(-2.0f * t + 2.0f, 2.0f) / 2.0f;
 	}
 
 	inline float easeInCubic(float t)
@@ -52,15 +52,15 @@ namespace MAD::Easings
 
 	inline float easeOutCubic(float t)
 	{
-		return 1 - std::pow(1 - t, 3);
+		return 1.0f - std::pow(1.0f - t, 3.0f);
 	}
 
 	inline float easeInOutCubic(float t)
 	{
 		if (t < 0.5f)
-			return 4 * t * t * t;
+			return 4.0f * t * t * t;
 		else
-			return 1 - std::pow(-2 * t + 2, 3) / 2;
+			return 1.0f - std::pow(-2.0f * t + 2.0f, 3.0f) / 2.0f;
 	}
 
 	inline float easeInQuart(float t)
@@ -70,15 +70,15 @@ namespace MAD::Easings
 
 	inline float easeOutQuart(float t)
 	{
-		return 1 - std::pow(1 - t, 4);
+		return 1.0f - std::pow(1.0f - t, 4.0f);
 	}
 
 	inline float easeInOutQuart(float t)
 	{
 		if (t < 0.5f)
-			return 8 * t * t * t * t;
+			return 8.0f * t * t * t * t;
 		else
-			return 1 - std::pow(-2 * t + 2, 4) / 2;
+			return 1.0f - std::pow(-2.0f * t + 2.0f, 4.0f) / 2.0f;
 	}
 
 	inline float easeInQuint(float t)
@@ -88,59 +88,59 @@ namespace MAD::Easings
 
 	inline float easeOutQuint(float t)
 	{
-		return 1 - std::pow(1 - t, 5);
+		return 1.0f - std::pow(1.0f - t, 5.0f);
 	}
 
 	inline float easeInOutQuint(float t)
 	{
 		if (t < 0.5f)
-			return 16 * t * t * t * t * t;
+			return 16.0f * t * t * t * t * t;
 		else
-			return 1 - std::pow(-2 * t + 2, 5) / 2;
+			return 1.0f - std::pow(-2.0f * t + 2.0f, 5.0f) / 2.0f;
 	}
 
 	inline float easeInExpo(float t)
 	{
-		return (t == 0) ? 0 : std::pow(2, 10 * t - 10);
+		return (t == 0.0f) ? 0.0f : std::pow(2.0f, 10.0f * t - 10.0f);
 	}
 
 	inline float easeOutExpo(float t)
 	{
-		return (t == 1) ? 1 : 1 - std::pow(2, -10 * t);
+		return (t == 1.0f) ? 1.0f : 1.0f - std::pow(2.0f, -10.0f * t);
 	}
 
 	inline float easeInOutExpo(float t)
 	{
-		if (t == 0 || t == 1)
+		if (t == 0.0f || t == 1.0f)
 			return t;
 		if (t < 0.5f)
-			return std::pow(2, 20 * t - 10) / 2;
+			return std::pow(2.0f, 20.0f * t - 10.0f) / 2.0f;
 		else
-			return (2 - std::pow(2, -20 * t + 10)) / 2;
+			return (2.0f - std::pow(2.0f, -20.0f * t + 10.0f)) / 2.0f;
 	}
 
 	inline float easeInCirc(float t)
 	{
-		return 1 - std::sqrt(1 - (t * t));
+		return 1.0f - std::sqrt(1.0f - (t * t));
 	}
 
 	inline float easeOutCirc(float t)
 	{
-		return std::sqrt(1 - std::pow(t - 1, 2));
+		return std::sqrt(1.0f - std::pow(t - 1.0f, 2.0f));
 	}
 
 	inline float easeInOutCirc(float t)
 	{
 		if (t < 0.5f)
-			return (1 - std::sqrt(1 - std::pow(2 * t, 2))) / 2;
+			return (1.0f - std::sqrt(1.0f - std::pow(2.0f * t, 2.0f))) / 2.0f;
 		else
-			return (std::sqrt(1 - std::pow(-2 * t + 2, 2)) + 1) / 2;
+			return (std::sqrt(1.0f - std::pow(-2.0f * t + 2.0f, 2.0f)) + 1.0f) / 2.0f;
 	}
 
 	inline float easeInBack(float t)
 	{
 		constexpr float c1 = 1.70158f;
-		constexpr float c3 = c1 + 1;
+		constexpr float c3 = c1 + 1.0f;
 
 		return c3 * t * t * t - c1 * t * t;
 	}
@@ -148,8 +148,8 @@ namespace MAD::Easings
 	inline float easeOutBack(float t)
 	{
 		constexpr float c1 = 1.70158f;
-		constexpr float c3 = c1 + 1;
-		return 1 + c3 * std::pow(t - 1, 3) + c1 * std::pow(t - 1, 2);
+		constexpr float c3 = c1 + 1.0f;
+		return 1.0f + c3 * std::pow(t - 1.0f, 3.0f) + c1 * std::pow(t - 1.0f, 2.0f);
 	}
 
 	inline float easeInOutBack(float t)
@@ -157,36 +157,36 @@ namespace MAD::Easings
 		constexpr float c1 = 1.70158f;
 		constexpr float c2 = c1 * 1.525f;
 		if (t < 0.5f)
-			return (std::pow(2 * t, 2) * ((c2 + 1) * 2 * t - c2)) / 2;
+			return (std::pow(2.0f * t, 2.0f) * ((c2 + 1.0f) * 2.0f * t - c2)) / 2.0f;
 		else
-			return (std::pow(2 * t - 2, 2) * ((c2 + 1) * (t * 2 - 2) + c2) + 2) / 2;
+			return (std::pow(2.0f * t - 2.0f, 2.0f) * ((c2 + 1.0f) * (t * 2.0f - 2.0f) + c2) + 2.0f) / 2.0f;
 	}
 
 
 	inline float easeInElastic(float t)
 	{
-		constexpr float c4 = (2 * M_PI) / 3;
+		constexpr float c4 = static_cast<float>((2.0f * M_PI) / 3.0f);
 
-		return (t == 0) 
-			? 0 
-			: (t == 1) 
-			? 1 
-			: -std::pow(2, 10 * t - 10) * std::sin((t * 10 - 10.75f) * c4);
+		return (t == 0.0f)
+			? 0.0f 
+			: (t == 1.0f)
+			? 1.0f 
+			: static_cast<float>( - std::pow(2.0f, 10.0f * t - 10.0f)) * static_cast<float>(std::sin((t * 10.0f - 10.75f) * c4));
 	}
 
 	inline float easeOutElastic(float t)
 	{
-		constexpr float c4 = (2 * M_PI) / 3;
-		return (t == 0) 
-			? 0 
-			: (t == 1) 
-			? 1 
-			: std::pow(2, -10 * t) * std::sin((t * 10 - 0.75f) * c4) + 1;
+		constexpr float c4 = static_cast<float>((2.0f * M_PI) / 3.0f);
+		return (t == 0.0f)
+			? 0.0f
+			: (t == 1.0f)
+			? 1.0f
+			: static_cast<float>(std::pow(2.0f, -10.0f * t) * std::sin((t * 10.0f - 0.75f) * c4) + 1.0f);
 	}
 
 	inline float easeInOutElastic(float t)
 	{
-		constexpr float c5 = (2.0f * M_PI) / 4.5f;
+		constexpr float c5 = static_cast<float>((2.0f * M_PI) / 4.5f);
 		if (t == 0 || t == 1.0f)
 			return t;
 		if (t < 0.5f)
@@ -227,15 +227,15 @@ namespace MAD::Easings
 
 	inline float easeInBounce(float t)
 	{
-		return 1 - MAD::Easings::easeOutBounce(1 - t);
+		return 1.0f - MAD::Easings::easeOutBounce(1.0f - t);
 	}
 
 
 	inline float easeInOutBounce(float t)
 	{
-		return t < 0.5f 
-			? (1 - MAD::Easings::easeOutBounce(1 - 2 * t)) / 2 
-			: (1 + MAD::Easings::easeOutBounce(2 * t - 1)) / 2;
+		return t < 0.5f
+			? (1.0f - MAD::Easings::easeOutBounce(1.0f - 2.0f * t)) / 2.0f
+			: (1.0f + MAD::Easings::easeOutBounce(2.0f * t - 1.0f)) / 2.0f;
 	}
